climate_policy_rag/requirements.txt
transformers==4.30.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.0.1
openai==0.27.0
langchain==0.0.217
